src/app/material/basic-elements/basic-elements.component.ts
import { Component, OnInit } from '@angular/core';
import { FormControl } from '@angular/forms';
import { TooltipPosition } from '@angular/material/tooltip';

@Component({
  selector: 'app-basic-elements',
  templateUrl: './basic-elements.component.html',
  styleUrls: ['./basic-elements.component.css']
})
export class BasicElementsComponent {

  divClicked = "clicknone";
  divHovered = "hovernone";
  positionOptions: TooltipPosition[] = ['below', 'above', 'left', 'right'];
  position = new FormControl(this.positionOptions[0]);

  onDivClick() {
    this.divClicked = "clicked";
  }

  onDivHover() {
    this.divHovered = "hovered";
  }
}